Set userid in nginx service container to uid of github runner user and restart service container
The hashfiles created by the ansible run during the integration tests belong to the github
runner user. As we do not adjust the permissions during the ansible run we have to set the
uid of the nginx user in the service container to the uid of the github runner. This
allows reading of this files by the nginx user after the service container got restarted
and the nginx process uses the new uid. Without this the validation would fail.