Add "local" provider for http-challenge (#30)
* Add "local" provider for http-challenge

Adds a local provider for http-challenge to allow http validation via a webserver
- split http-challenge into s3 and local playbooks which get included when letsencrypt_http_provider is set accordingly
- add letsencrypt_local_validation_path variable to defaults to work out of the box
- update documentation with new local provider for http-challenge
- update examples with letsencrypt_http_provider variable

* Update roles/letsencrypt/README.md

Co-authored-by: Sebastian Gumprich <rndmh3ro@users.noreply.github.com>

* update README to silence editor linter

- MD022 [blanks around headings]
- MD041 [first line heading]
- MD001 [heading increment]

* Add integration test for local http provider

* Start nginx as service to test http-challenge

Expose nginx on port 80 to allow pebble to validate over http.
Overwrite the vhost path with a volume as we will use it for the hashfiles.
Overwrite default path for local validation for use in integration test

* Fix syntax for mounting volumes in nginx service container

* Adjust naming of files and tasknames to distinguish between tests

Changed the naming of the resulting files to better distinguish between the different tests and also changed the tasknames to make it more clear what is actualy done

* Change certificate names and domains

Changed certificate names and domains to ensure a new certificate is realy issued and the whole process runs

* Correct taskname

* Install python-boto3 package only when s3 provider is used

* Change exposed port to 5002

5002 is the default validation port of pebble for HTTP-01 challenge

* Create directory for challenge/hash files or else playbook fails

The variable "challenge['challenge_data'][item]['http-01']['resource']" looks like this .well-known/acme-challenge/hashfile324324
Therefore it is a requirement to have this directories in place

* Add acme_directory_validate_certs variable
This adds the posibility to disable the validation of acme directory certs and run integration tests against the pebble container via http challenge


* Updated documentation with missing validate_certs var

* Add newline to silence markup linter

* Set userid in nginx service container to uid of github runner user and restart service container

The hashfiles created by the ansible run during the integration tests belong to the github
runner user. As we do not adjust the permissions during the ansible run we have to set the
uid of the nginx user in the service container to the uid of the github runner. This
allows reading of this files by the nginx user after the service container got restarted
and the nginx process uses the new uid. Without this the validation would fail.

* Adjust local validation path in test

Co-authored-by: Andreas Hering <andreas.hering@t-systems.com>
Co-authored-by: Sebastian Gumprich <rndmh3ro@users.noreply.github.com>