Rename Role and unify variables
To avoid trademark issues with LE we decided to rename the collection
(issue 43). Thus leading to a breaking change.
Variablenames where not consistent in their naming. They were unified in
conjunction of the collection renaming (issue 44) to better match the
name of the collection/role. As the renaming alrdy is a breaking change
we also changed this with this release.
Added some more tags to the collection/role to be found better in galaxy
search