Set role name  & unify variables once more
Revert some unify changes as they were not neccessary for items of a
list.
Set role name to acme_letsencrypt to better match the naming of collection
and role and thus allowing clearer naming of variables.
Unify variables to a better schema with following format:
collectionname_rolename(_provider)?_variablename
Fixed some errors which got in during search and replace of variables.