Rename collection | simplify provider selection | unify variables  (#46)
* Simplify challenge provider selection

Change used variable for challenge provider selection to
letsencrypt_*_provider.
Change preconditions check to check if letsencrypt_*_provider variable
is set.
Remove letsencrypt_do_*_challenge variables as they are not used
anymore.
Adjust documentation.

* Fix variable check of provider

* Rename Role and unify variables

To avoid trademark issues with LE we decided to rename the collection
(issue 43). Thus leading to a breaking change.
Variablenames where not consistent in their naming. They were unified in
conjunction of the collection renaming (issue 44) to better match the
name of the collection/role. As the renaming alrdy is a breaking change
we also changed this with this release.
Added some more tags to the collection/role to be found better in galaxy
search

* Adjust check for a set challenge provider

* Change included role in test integration back to letsencrypt as their name did not change

* Update documentation

* Set role name  & unify variables once more

Revert some unify changes as they were not neccessary for items of a
list.
Set role name to acme_letsencrypt to better match the naming of collection
and role and thus allowing clearer naming of variables.
Unify variables to a better schema with following format:
collectionname_rolename(_provider)?_variablename
Fixed some errors which got in during search and replace of variables.

* Fix naming of list elements when domain list is used

Updated documentation indentation of tables

* Fix git dir path in workflows

* Fix module name (got replaced during search and replace -.-)

* Rename role folder to acme_letsencrypt + adjust rolename in integration test

* Rename rolename in integrations test to new role name

* Skip ansible lint rule 602 as there is no other elegant not fiddly way to check for an empty string

* Add mode to creation of challenge directory to silence linter

* Ajust documentation to new naming

* Rename madness part 2

Rename role to acme.
Rename variables to match new role name.
Move set of challenge provider to preconditions.yml to have a clean
main.yml with only includes.
Documentation updates to descriptions, indentations, readme naming.

* Empty commit to trigger actions after renaming repo

* Update docs/role-acme.md

Co-authored-by: Sebastian Gumprich <rndmh3ro@users.noreply.github.com>

Co-authored-by: Andreas Hering <andreas.hering@t-systems.com>
Co-authored-by: Sebastian Gumprich <rndmh3ro@users.noreply.github.com>