Lookup ZoneID and fix challenge record format.
I had to make some adjustments to get this to work for Hetzner DNS. To add a record you need a zone ID, and the acme challenge TXT records need to include the sub domain. If you have a wildcard this needs to be stripped also. Similar to your azure module.