Fixing the half-validated bug
On certificates with multiple subject names (subject_alt_name), there is the possibility that a validation goes halfway through.

In such a case the filter_challenges filter plugin failed because it only searched in the authorizations field  when challenge_data field was empty. In a half-way validation the challenge_data will carry the tokens of the pending validations and therefore the find_challenges() fails on a expected_not_found validation.